requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
matplotlib>=3.8
PyYAML>=6.0
openpyxl>=3.1
pyarrow>=15.0
joblib>=1.3
holidays>=0.93
